Split installer into individually executable scripts
